refactor(credits, payments): move logic to models and add more tests
- Added additional tests for the 'credits' app to improve coverage
- Refactored 'credits' and 'payments' serializers by moving logic to
  their respective models.
