+++ .coveragerc
@@ -10,4 +10,5 @@ include =
     products/*
     users/*
     clients/*
+    credits/*
     utils/*
