+++ creditoapp/urls.py
@@ -32,7 +32,6 @@ router.register(r'product-types', ProductTypeViewSet)
 router.register(r'products', ProductViewSet)
 router.register(r'credits', CreditViewSet)
 router.register(r'payments', PaymentViewSet)
-#router.register(r'interest-rates', InterestRateViewSet)
 router.register(r'users', UserViewSet)
 router.register(r'clients', ClientViewSet)
